Fix var scoping
With scripts being compiled into their own closures we have to be a
little more careful about implicit global scoping. These undeclared vars
were getting hung on window, now we're being more explicit about which
ones are used internally to the script and which ones really exist on
window for the purpose of inline clikc handlers.